feat: Implement controls, footer, and timeline components with translation support
